Add OpenRefine project
Input file:

    join -t $'\t' names biographical > openrefine-in

But it turns out the biographical information doesn’t help at all with
automatic reconciliation (it lowers the confidence so that no
auto-matches remain), so I just reconciled on the name by its own. (But
the biographical information will still be useful for the manual
resolutions necessary.)

After reconciliation with Wikidata, 3197 entries matched successfully,
which leaves 876 to be matched manually.