Add AWK include file with some utility functions
The dump specifies dates in the DD.MM.YYYY format common in Germany, but
we’re more likely to need them in ISO 6801 YYYY-MM-DD format, more
specifically in QuickStatements +YYYY-MM-DDT00:00:00/11 format. This
file contains some functions to convert to these formats. (More
functions and variables may be added in the future, so the file name is
kept very generic for now.)