Add Bundestag ID to references (as catalog code)
Andrew Gray pointed out that it would be nice to record the ID in the
reference as well, so other people can verify the reference without our
ID file.

This requires us to emit a string value, which gets a bit ugly –
QuickStatements parses the CSV using PHP’s fgetcsv() function, and
convincing that function to emit a column wrapped in double quotes (so
QuickStatements interprets it as a string) apparently requires *three*
double quotes in the original CSV. But we were going to run into that
problem sooner or later (I’m sure we’ll want to import some kind of
strings eventually), so it doesn’t hurt to have dealt with it now.

QuickStatements code still generated with the same command:

    join -t$'\t' IDs biographical | head -n50 | ./dob+dod.awk >| dob+dod-1.qs.csv