Add JavaScript code to add edit summary to QuickStatements
This is a fairly ugly hack, but it lets you edit with QuickStatements
while setting a custom edit summary. Just paste it into your browser’s
web console and it should work, tweaking the client-side website
accordingly. (Fortunately, no server-side changes are necessary, or we’d
be out of luck.) Based on [1]; see also Magnus’ comments in [2].

The run function is a slightly edited version of the original from
QuickStatements’ source code, which is not currently published under any
particular license as far as I can tell.

[1]: https://gist.github.com/lucaswerkmeister/d6a6a64060dba38c4bd09dd939af72d8
[2]: https://twitter.com/LucasWerkmeistr/status/989875106963972096