Skip entries without date of death for now
QuickStatements doesn’t do well with empty fields in the CSV import (it
results in a PHP error inside the response that’s supposed to be JSON,
so the website fails to parse the status and completely stops
processing), so the easiest way to proceed for now will be to import
dates of birth/death separately for people with a date of death and
people without one.

QuickStatements code still generated with the same command:

    join -t$'\t' IDs biographical | head -n50 | ./dob+dod.awk >| dob+dod-1.qs.csv

But since we’re filtering inside AWK now, the final file has less than
50 lines, so we need to take care to remember how many input lines to
skip! If I’m not mistaken, the next batch should be resumed with:

    join -t$'\t' IDs biographical | tail -n+51 | head -nTODO | ./dob+dod.awk >| dob+dod-1.qs.csv

where the TODO limit depends on how fast the first batch goes.