Add second QuickStatements batch: 500 rows
The first batch of 50 input rows (though some were discarded due to
lacking a date of death and a row does not exactly correspond to a
certain number of edits) finished in *roughly* one minute, so a batch of
500 input rows should easily finish tonight before I shut down the
computer and go to bed.

Generated with:

    join -t$'\t' IDs biographical | tail -n+51 | head -n500 | ./dob+dod.awk >| dob+dod-2.qs.csv

I manually verified that this batch resumes right where the previous one
left off – no input rows should be skipped.

Next batch with (I think):

    join -t$'\t' IDs biographical | tail -n+551 | head -nTODO | ./dob+dod.awk >| dob+dod-3.qs.csv