Add third QuickStatements batch: 1000 rows
The second batch of 500 input rows completed in just under 16 minutes,
so there should be enough time for a 1000-row batch tonight.

Generated with:

   join -t$'\t' IDs biographical | tail -n+551 | head -n1000 | ./dob+dod.awk >| dob+dod-3.qs.csv

I verified that no input rows with date of death were skipped between
batches 2 and 3 (though there are a few lines without date of death
right around the threshold between the batches).

Next batch with:

    join -t$'\t' IDs biographical | tail -n+1551 | head -nTODO | ./dob+dod.awk >| dob+dod-4.qs.csv