Add seventh QuickStatements batch: 2000 rows
Now we start with the MdBs without a date of death yet. The AWK script
no longer emits columns or column headers for the date of death, and
skips all entries that have one. Since the deceased MdBs outnumber the
living ones (especially in the beginning of the dump), we can start with
quite a large batch of input rows and it won’t actually translate to
*too* many output rows.

Generated with:

    join -t$'\t' IDs biographical | head -n2000 | ./dob+dod.awk >| dob+dod-7.qs.csv

Next batch with:

    join -t$'\t' IDs biographical | tail -n+2001 | head -nTODO | ./dob+dod.awk >| dob+dod-8.qs.csv